+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Version - 0]
 
+## improve [v0.46.5] 2024-1-19
+
+- target spray order
+
 ## improve [v0.46.4] 2024-1-18
 
 - added Equal Price